requirements.txt
datasets>=1.6.1
evaluate>=0.3.0
hydra-core>=1.2.0
numpy>=1.19.5
python-dotenv>=0.21.0
pytorch-lightning>=1.7.7
torch>=1.12.1
tqdm>=4.64.0
transformers>=4.23.1
sacremoses>=0.0.53
wandb>=0.13.5
